Blacklisting functional
This commit implements blacklisting and dead letter queue delivery (a mocked implementation of dead letter delivery where it just logs the event message).

There are a number of things that feel a bit hackey. Especially as it relates to getting the event sequence identifier. My code makes the assumption
that it will be the aggregate identifier when the code doesn't have access to an event message which is required to get the value from the Sequencing policy.

Also, this commit implements a strategy where the event handler has to opt-in to acting on blacklisted event sequences. I had originally intended to make
blacklisting and the dead letter queue the only way to handle failure, but I couldn't seem to get what I needed from the interceptor. So, instead of
doing it automatically, I set up a deal where the event handler will trigger a `EventQuarantinedException` when the event sequence is blacklisted. This
thereby enables the event handler to decide if failed events can be skipped and subsequent events processed, or if it should use the dead letter queue...