naming refactor plus responding to EventSequenceBlacklisted
Firstly, I refactored the naming of the event handling failure and blacklist related
components and cleaned up the implementation some.

Also, I added an event EventSequenceBlacklisted which is raised each time a BlacklistedEventSequence
is created. This gives eventhandling components a chance to respond the blacklisting failure.

I've add to the UserProfileProjector an event handler to handle EventSequenceBlacklisted by marking
the profile with an error status.